DRY up logic to generate legacy-compatible array APIs (#465)
